Tests/LibWeb/Text/expected/Fetch/fetch-later-multiple-iframe-unload.txt

all cookies observed after iframe unload: true
cookies observed before iframe unload: false
activated count before iframe unload: 0
PASS
